wm-actions-signal
signal names

to squash later

same

right
