+++ plugins/meson.build
@@ -6,3 +6,4 @@ subdir('cube')
 subdir('wobbly')
 subdir('blur')
 subdir('tile')
+subdir('wm-actions')
