+++ plugins/wm-actions/wm-actions-signals.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <wayfire/signal-definitions.hpp>
+
+/**
+ * name: wm-actions-toggle-above
+ * on: output
+ * when: Emitted whenever some entity requests that the view's above state
+ *   is supposed to change.
+ */
+using wm_actions_toggle_above = wf::_view_signal;
+
+/**
+ * name: wm-actions-toggle-above-changed
+ * on: output
+ * when: Emitted whenever a views above layer has been changed.
+ */
+using wm_actions_above_changed = wf::_view_signal;
